+++ src/components/Forms/InteractionModalForm/InteractionModalForm.tsx
@@ -8,6 +8,7 @@ import {
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Transaction } from '@multiversx/sdk-core/out';
 import { DECIMALS } from '@multiversx/sdk-dapp/constants/index';
+import { getEgldLabel } from '@multiversx/sdk-dapp/utils/network/getEgldLabel';
 import { parseAmount } from '@multiversx/sdk-dapp/utils/operations/parseAmount';
 import { stringIsFloat } from '@multiversx/sdk-dapp/utils/validation/stringIsFloat';
 import BigNumber from 'bignumber.js';
@@ -61,7 +62,7 @@ export const InteractionModalForm = (props: InteractionModalFormUIType) => {
     icons,
     customClassNames
   } = useSCExplorerContext();
-  const getTransactionCost = useGetTransactionCost();
+  const getTransactionCost = useGetTransactionCost({ isDeploy });
   const { deployedContractDetails, abiRegistry, contractAddress } =
     smartContract ?? {};
   const {
@@ -87,8 +88,8 @@ export const InteractionModalForm = (props: InteractionModalFormUIType) => {
   } = accountInfo;
   const [selectedToken, setSelectedToken] = useState<SelectOptionType>();
   const [isTxCostLoading, setIsTxCostLoading] = useState(false);
-  const [simulatedTxGasLimit, setSimulatedTxGasLimit] = useState<
-    number | undefined
+  const [hasVerifiedGasLimit, setHasVerifiedTxGasLimit] = useState<
+    boolean | undefined
   >();
 
   const metadataOptionsInitialValues =
@@ -126,6 +127,8 @@ export const InteractionModalForm = (props: InteractionModalFormUIType) => {
     egldBalance
   });
 
+  const egldLabel = getEgldLabel();
+
   const validationSchema = object({
     gasLimit: string()
       .required('Required')
@@ -150,6 +153,16 @@ export const InteractionModalForm = (props: InteractionModalFormUIType) => {
                 selectedToken.token.balance ?? '0'
               );
               return bnBalance.comparedTo(bnAmount) >= 0;
+            }
+            return true;
+          })
+          .test('hasValidValue', 'ESDToken amount cannot be 0', (value) => {
+            if (
+              value &&
+              selectedToken?.token?.identifier &&
+              selectedToken.token.identifier !== egldLabel
+            ) {
+              return new BigNumber(value.toString()).isGreaterThan(0);
             }
             return true;
           }),
@@ -210,10 +223,8 @@ export const InteractionModalForm = (props: InteractionModalFormUIType) => {
     }
 
     if (transaction) {
-      const gasLimit = await getTransactionCost(transaction);
-      if (gasLimit) {
-        return gasLimit;
-      }
+      const transactionGasLimit = await getTransactionCost(transaction);
+      return transactionGasLimit;
     }
 
     return;
@@ -277,21 +288,22 @@ export const InteractionModalForm = (props: InteractionModalFormUIType) => {
 
           const fetchData = async () => {
             setIsTxCostLoading(true);
-            const gasLimit = await getTransactionCostDetails({
+            const transactionGasLimit = await getTransactionCostDetails({
               tokens
             });
-            if (gasLimit) {
+            if (transactionGasLimit) {
+              const { gasLimit, isVerified } = transactionGasLimit;
               setFieldValue(
                 InteractionModalFormikFieldsEnum.gasLimit,
                 gasLimit
               );
-              setSimulatedTxGasLimit(gasLimit);
+              setHasVerifiedTxGasLimit(isVerified);
             } else {
               setFieldValue(
                 InteractionModalFormikFieldsEnum.gasLimit,
                 defaultGasLimit
               );
-              setSimulatedTxGasLimit(0);
+              setHasVerifiedTxGasLimit(false);
             }
             setIsTxCostLoading(false);
           };
@@ -299,11 +311,11 @@ export const InteractionModalForm = (props: InteractionModalFormUIType) => {
           if (
             validTokens &&
             !isGasValueTouched &&
-            simulatedTxGasLimit === undefined
+            hasVerifiedGasLimit === undefined
           ) {
             fetchData().catch(console.error);
           }
-        }, [values, simulatedTxGasLimit]);
+        }, [values, hasVerifiedGasLimit]);
 
         return (
           <Form
@@ -391,7 +403,7 @@ export const InteractionModalForm = (props: InteractionModalFormUIType) => {
                   )}
                 />
                 {Boolean(
-                  (isTxCostLoading || simulatedTxGasLimit) && !isGasValueTouched
+                  (isTxCostLoading || hasVerifiedGasLimit) && !isGasValueTouched
                 ) && (
                   <div
                     className={classNames(
@@ -407,7 +419,7 @@ export const InteractionModalForm = (props: InteractionModalFormUIType) => {
                       />
                     ) : (
                       <>
-                        {simulatedTxGasLimit && (
+                        {hasVerifiedGasLimit && (
                           <FontAwesomeIcon icon={faCircleCheck} />
                         )}
                       </>