+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Updated dependencies, Fix upgrade endpoint, WebWallet login using CrossWindow](https://github.com/multiversx/mx-sdk-dapp-sc-explorer/pull/20)
+
 ## [[0.0.1-beta.7]](https://github.com/multiversx/mx-sdk-dapp-sc-explorer/pull/19)] - 2024-03-14
 
 - [Updated dependencies in order to fix ABI parsing](https://github.com/multiversx/mx-sdk-dapp-sc-explorer/pull/19)